+++ PTRANS/src/host/transpose_benchmark.cpp
@@ -111,10 +111,10 @@ transpose::TransposeBenchmark::validateOutputAndPrintError(transpose::TransposeD
     // exchange the data using MPI depending on the chosen distribution scheme
     dataHandler->exchangeData(data);
 
-    int block_offset = executionSettings->programSettings->blockSize * executionSettings->programSettings->blockSize;
-    for (int b = 0; b < data.numBlocks; b++) {
-        for (int i = 0; i < executionSettings->programSettings->blockSize; i++) {
-            for (int j = 0; j < executionSettings->programSettings->blockSize; j++) {
+    size_t block_offset = executionSettings->programSettings->blockSize * executionSettings->programSettings->blockSize;
+    for (size_t b = 0; b < data.numBlocks; b++) {
+        for (size_t i = 0; i < executionSettings->programSettings->blockSize; i++) {
+            for (size_t j = 0; j < executionSettings->programSettings->blockSize; j++) {
                 data.A[b * block_offset + j * executionSettings->programSettings->blockSize + i] -= (data.result[b * block_offset + i * executionSettings->programSettings->blockSize + j] 
                                                                             - data.B[b * block_offset + i * executionSettings->programSettings->blockSize + j]);
             }
@@ -122,7 +122,7 @@ transpose::TransposeBenchmark::validateOutputAndPrintError(transpose::TransposeD
     }
 
     double max_error = 0.0;
-    for (int i = 0; i < executionSettings->programSettings->blockSize * executionSettings->programSettings->blockSize * data.numBlocks; i++) {
+    for (size_t i = 0; i < executionSettings->programSettings->blockSize * executionSettings->programSettings->blockSize * data.numBlocks; i++) {
         max_error = std::max(fabs(data.A[i]), max_error);
     }
